extract `gather_leaves` function
