+++ src/parse/reduction/reduction.c
@@ -1,4 +1,5 @@
 #include "reduction.h"
+#include "reduction_internals.h"
 #include "log/log.h"
 #include "parse/t_command/t_command.h"
 #include "parse/t_symbol/t_symbol.h"
@@ -28,36 +29,6 @@ t_symbol* find_symbol(t_symbol* root, t_symbol_kind kind)
 	}
 
     return NULL;
-}
-
-static void recurse(t_symbol* root, t_token_list** leaves_p, t_symbol_stack** visited)
-{
-    ss_push(visited, root);
-    if (root->kind == TERMINAL)
-    {
-        tkl_push_back(leaves_p, root->token);
-    }
-    else
-    {
-		for (size_t i = 0; i < root->production->len; i++)
-		{
-		    t_symbol* candidate = &root->production->data[i];
-		    if (!ss_contains(*visited, candidate))
-				recurse(candidate, leaves_p, visited);
-		}
-    }
-}
-
-t_token_list *gather_leaves(t_symbol* root)
-{
-	t_token_list* leaves;
-    t_symbol_stack* visited;
-
-	leaves = NULL;
-	visited = NULL;
-    recurse(root, &leaves, &visited);
-	//ss_clear(&visited);
-	return (leaves);
 }
 
 #include <stdlib.h>
